Minor fixes

+++ nefics/modules/honeypot.py
@@ -5,6 +5,7 @@ import subprocess
 import os
 from time import sleep
 from datetime import datetime
+from typing import Union, Callable
 # NEFICS imports
 from nefics.modules.devicebase import IEDBase, DeviceHandler, ProtocolListener, LOG_PRIO
 from nefics.protos import http, modbus
@@ -30,7 +31,7 @@ class HoneyHandler(DeviceHandler):
         check_honey = subprocess.call(['which', 'honeyd'], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, shell=False)
         if check_honey != 0:
             raise FileNotFoundError(2, 'Cannot find executable', 'honeyd')
-        self._process = None
+        self._process : Union[subprocess.Popen, None] = None
     
     def _respawn(self):
         if isinstance(self._process, subprocess.Popen):
@@ -59,16 +60,17 @@ class HoneyHandler(DeviceHandler):
                 # Process terminated, respawning
                 self._respawn()
             sleep(5)
-        self._process.terminate()
-        self._process.wait(20)
+        if isinstance(self._process, subprocess.Popen):
+            self._process.terminate()
+            self._process.wait(20)
 
 class PLCDevice(IEDBase):
 
     def __init__(self, guid: int, neighbors_in: list = ..., neighbors_out: list = ..., **kwargs):
         super().__init__(guid, neighbors_in, neighbors_out, **kwargs)
-        self._html : str = kwargs['html'] if 'html' in kwargs.keys() and isinstance(kwargs['html'], str) else None
+        self._html : Union[str, None] = kwargs['html'] if 'html' in kwargs.keys() and isinstance(kwargs['html'], str) else None
         self._httpsrv = kwargs['httpsrv'] if 'httpsrv' in kwargs.keys() and isinstance(kwargs['httpsrv'], str) else None
-        self._protocols = kwargs['protos'] if 'protos' in kwargs.keys() and isinstance(kwargs['protos'], list) and all(isinstance(x, str) for x in kwargs['protos']) else None
+        self._protocols : Union[list[str], None] = kwargs['protos'] if 'protos' in kwargs.keys() and isinstance(kwargs['protos'], list) and all(isinstance(x, str) for x in kwargs['protos']) else None
     
     @property
     def httpsrv_header(self) -> str:
@@ -79,7 +81,7 @@ class PLCDevice(IEDBase):
         return self._html if self._html is not None else ''
     
     @property
-    def protocols(self) -> list[str]:
+    def protocols(self) -> Union[list[str], None]:
         return self._protocols
     
     def __str__(self) -> str:
@@ -117,14 +119,11 @@ class PLCHandler(DeviceHandler):
 
     def _start_http(self):
         try:
-            http_args : dict[str,str] = {}
-            http_args['server_header'] = self._device.httpsrv_header
             hpath = self._device.html_path
             assert len(hpath)
             assert os.path.exists(hpath)
             assert os.path.isdir(hpath)
-            http_args['static_dir'] = hpath
-            listener : ProtocolListener = http.HTTPListener(**http_args)
+            listener : ProtocolListener = http.HTTPListener(server_header=self._device.httpsrv_header, static_dir=hpath)
             self._protocols['http'] = listener
             listener.start()
         except AssertionError:
@@ -138,7 +137,7 @@ class PLCHandler(DeviceHandler):
     def run(self):
         self._device.start()
         if self._device.protocols is not None:
-            phandlers : dict[str, function]= {
+            phandlers : dict[str, Callable]= {
                 'http' : self._start_http,
                 'modbus' : self._start_modbus
             }
